requirements.txt
nuscenes-devkit
numpy
pandas
notebook
seaborn
pyYAML
tqdm
ensure==1.0.2
python-box==6.0.2
joblib
types-PyYAML
Flask
Flask-Cors
-e .
